Added basic routing, casting, and API.
The API may be temporary and is acting as a placeholder for testing
purposes. The casting will depend on methods that will only be
accessible when the solution is run on the target Raspberry Pi, so--in
this commit, at least--is largely untestable. Also, a .gitignore for my
virtualenvironment was added. Added dependency for Flask and requests